refactor!: Let users and middleware change all parts of a request
A problem with letting middleware work on a requests.Request is that
transport options are not set on the Request object. Transport options,
such as timeout and cert, are given directly as keyword arguments to
session.send(). Hence, middleware had no way to set these, making it
impossible to implement certain behaviors.

This called for a design decision. One way is to keep them separate and
introduce send_kwargs to process_request (or possibly box the Request
and other options in a structure we define ourselves). This solves the
immediate problem for middleware, however, I'd like for users to be able
to also make configuration changes on an per-call basis. This is useful
for debugging and that one request you have that takes much longer than
all others. Another way is to use requests.request instead, which
collects all arguments for both a Request and .send() in one method,
which allows us to gather all argument destined for requests in
double-star parameters. This is what I ultimately went for because it is
a very thin (and hence flexible) way to wrap a request that also allows
us to merge arguments set at different levels. The priority goes, from
lowest to highest: endpoint options -> middleware -> call-site options.

The price we pay is that it is less obvious for implementers of
middleware what they can add or alter; the parameter is just a dict, not
a Request object with a documented parameter set.

BREAKING CHANGE: middleware that implements process_request must change
the signature to accept only **requests_kwargs, which will ultimately be
passed as kwargs to a requests.request call.