fix: requests_kwargs is an optional keyword argument to post endpoints
I.e., use a None default value.

Post endpoints must wrap their requests arguments in this dict because
the normal behaviour of post is to pass the unknown args and kwargs to
the wrapped function (f). In turn, f proceeds to build a json body from
its arguments. Hence, we need separation between request kwargs and
arguments for f.