refactor: split into test and release, manual release workflow
We split the workflows because tests should also run on
pull_request, but the release workflow only when a new tag
is pushed.

Switch to a manual trigger for the release workflow to be able to test
it. Needs to be merged to main before it can be triggered on
GitHub, however. I don't use the gh CLI, so let's do it the awkward
way.