ci: more pre-commit hooks for sanity; use in CI
Prevent fixes such as 3ece147b9a65faf8851e92732bf5ea732d27a299 from having to be applied again.

Use the configuration file hooks in the CI for the same reasons as we
run ruff check. We run ruff directly because the ruff's pre-commit hook
is configured to also fix the errors, which we don't want in the CI.